pyproject.toml
[project]
name = "agent-dqe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.10.5",
    "google-adk>=1.7.0",
]
